requirements.txt
django==1.11.29
pytz==2017.2
psycopg2-binary==2.8.*
python-dotenv==0.19.*
django-environ==0.8.*